Add files via upload
The blue laser will be for the player. Yellow/orange will be for the altered M-space enemy ships, and green will be for bugs. The first frame of each is for a charge-up effect, but could be put at the end of a beam when it hits something as well. There's two frames of each beam so that we can optionally have a little animation on the beams.